[box] improved how the enable/disable button is added.
[treeview]
changed insertItem() to take a reference instead of a pointer.
changed matchItemText() to use DcxListHelper::matchItemText()
changed parsePath() to take a reference TString object instead of a pointer.

[general]
added some static_cast's
added TiXmlElement *toXml()
changed toXml() to properly save the control styles too.
added DcxIconSizes enum to set standard easy to read icon sizes.
added NumToIconSize() helper function to allow easy conversion of any number to the enum DcxIconSizes

+++ Classes/dcxmdialog.cpp
@@ -49,7 +49,7 @@ DcxMDialog::DcxMDialog(const HWND cHwnd, const HWND pHwnd, const UINT ID, DcxDia
 	ShowWindow(m_Hwnd, SW_SHOWNOACTIVATE);
 	UpdateWindow(m_Hwnd);
 
-	this->m_OrigID = SetWindowLong(m_Hwnd, GWL_ID, ID);
+	this->m_OrigID = static_cast<UINT>(SetWindowLong(m_Hwnd, GWL_ID, static_cast<LONG>(ID)));
 
 	this->registreDefaultWindowProc();
 	SetProp(m_Hwnd, TEXT("dcx_cthis"), (HANDLE) this);
@@ -73,7 +73,7 @@ DcxMDialog::~DcxMDialog( ) {
 		if ( !bHide )
 			ShowWindow( m_Hwnd, SW_HIDE );
 
-		SetWindowLong( m_Hwnd, GWL_ID, this->m_OrigID );
+		SetWindowLong( m_Hwnd, GWL_ID, static_cast<LONG>(this->m_OrigID ));
 		//this->removeStyle(WS_CHILD);
 		//this->addStyle(WS_POPUP);
 		//SetParent( m_Hwnd, nullptr );
